[refactor] Change name of Permission component
